\name{DaxCacNik}
\docType{data}
\alias{DaxCacNik}

\title{Log-returns of daily indices of stock markets in Frankfurt, Paris and Tokio}

\description{
  The matrix \code{DaxCacNik} contains daily observations of the
  hundredfold log-returns of daily indices of stock markets in Frankfurt (DAX), 
  Paris (CAC40) and Tokyo (NIKKEI), from 10 October 1991 until 30 December 1997 
  (a total of 1627 days). The stock market data is freely available at 
  \url{http://robjhyndman.com/tsdldata/data/FVD1.dat}.
}

\usage{data(DaxCacNik)}

\references{Fioruci, J.A., Ehlers, R.S. Andrade Filho, M.G. 
	\emph{Bayesian multivariate GARCH models with dynamic correlations 
	and asymmetric error distributions}, Journal of Applied Statistics, 
	41(2), 320--331, 2014. \url{http://dx.doi.org/10.1080/02664763.2013.839635}.
}

\author{
	Jose Augusto Fioruci, Ricardo Sandes Ehlers and Francisco Louzada
}

\keyword{datasets}




